refactor: share AVX2 parity test helpers
