+++ crates/bitnet-kernels/tests/avx2_activation_reduction_parity_tests.rs
@@ -19,6 +19,10 @@ use bitnet_kernels::cpu::simd_reduction::{
     simd_argmax, simd_argmin, simd_horizontal_max, simd_horizontal_min, simd_horizontal_sum,
 };
 
+#[path = "common/avx2_parity.rs"]
+mod avx2_parity;
+use avx2_parity::{assert_vec_parity, close, pseudo_rand};
+
 // ── Tolerance constants ────────────────────────────────────────────────
 //
 // Activation functions use approximated exp/tanh internally so we allow
@@ -33,38 +37,6 @@ const ELEM_REL_TOL: f32 = 1e-4;
 const REDUCTION_ABS_TOL: f32 = 1e-5;
 /// Relative tolerance for reduction operations.
 const REDUCTION_REL_TOL: f32 = 1e-5;
-
-// ── Helpers ────────────────────────────────────────────────────────────
-
-fn close(a: f32, b: f32, abs_tol: f32, rel_tol: f32) -> bool {
-    let diff = (a - b).abs();
-    diff <= abs_tol || diff <= rel_tol * a.abs().max(b.abs())
-}
-
-fn assert_vec_parity(actual: &[f32], expected: &[f32], abs_tol: f32, rel_tol: f32, ctx: &str) {
-    assert_eq!(actual.len(), expected.len(), "{ctx}: length mismatch");
-    for (i, (&a, &e)) in actual.iter().zip(expected.iter()).enumerate() {
-        assert!(
-            close(a, e, abs_tol, rel_tol),
-            "{ctx}[{i}]: scalar={e}, dispatched={a} (diff={}, abs_tol={abs_tol}, rel_tol={rel_tol})",
-            (a - e).abs()
-        );
-    }
-}
-
-/// Deterministic pseudo-random f32 values in [-1, 1] for reproducibility.
-fn pseudo_rand(len: usize, seed: u64) -> Vec<f32> {
-    let mut state = seed;
-    (0..len)
-        .map(|_| {
-            // xorshift64
-            state ^= state << 13;
-            state ^= state >> 7;
-            state ^= state << 17;
-            (state as f32 / u64::MAX as f32) * 2.0 - 1.0
-        })
-        .collect()
-}
 
 // ── Scalar reference implementations ──────────────────────────────────
