+++ crates/bitnet-kernels/tests/avx2_elementwise_parity_tests.rs
@@ -16,6 +16,10 @@
 use bitnet_kernels::cpu::elementwise_ops;
 use std::f32::consts::PI;
 
+#[path = "common/avx2_parity.rs"]
+mod avx2_parity;
+use avx2_parity::{assert_vec_parity, pseudo_rand};
+
 // ── Tolerance constants ────────────────────────────────────────────────
 //
 // Arithmetic ops (add/sub/mul/div/fma) should be bit-exact or within
@@ -31,38 +35,6 @@ const ARITH_REL_TOL: f32 = 1e-6;
 const TRANS_ABS_TOL: f32 = 1e-5;
 /// Relative tolerance for transcendental / activation kernels.
 const TRANS_REL_TOL: f32 = 1e-4;
-
-// ── Helpers ────────────────────────────────────────────────────────────
-
-fn close(a: f32, b: f32, abs_tol: f32, rel_tol: f32) -> bool {
-    let diff = (a - b).abs();
-    diff <= abs_tol || diff <= rel_tol * a.abs().max(b.abs())
-}
-
-fn assert_vec_parity(actual: &[f32], expected: &[f32], abs_tol: f32, rel_tol: f32, ctx: &str) {
-    assert_eq!(actual.len(), expected.len(), "{ctx}: length mismatch");
-    for (i, (&a, &e)) in actual.iter().zip(expected.iter()).enumerate() {
-        assert!(
-            close(a, e, abs_tol, rel_tol),
-            "{ctx}[{i}]: scalar={e}, dispatched={a} (diff={}, abs_tol={abs_tol}, rel_tol={rel_tol})",
-            (a - e).abs()
-        );
-    }
-}
-
-/// Deterministic pseudo-random f32 values in [-1, 1] for reproducibility.
-fn pseudo_rand(len: usize, seed: u64) -> Vec<f32> {
-    let mut state = seed;
-    (0..len)
-        .map(|_| {
-            // xorshift64
-            state ^= state << 13;
-            state ^= state >> 7;
-            state ^= state << 17;
-            (state as f32 / u64::MAX as f32) * 2.0 - 1.0
-        })
-        .collect()
-}
 
 /// Test lengths: 63 (non-8-aligned, exercises scalar tail) and 128 (8-aligned).
 const LENGTHS: [usize; 2] = [63, 128];
