+++ crates/bitnet-kernels/tests/avx2_embedding_pooling_parity_tests.rs
@@ -12,6 +12,10 @@ use bitnet_kernels::cpu::pooling::{
     PoolConfig, PoolType, avg_pool1d, avg_pool1d_avx2, max_pool1d, max_pool1d_avx2,
 };
 
+#[path = "common/avx2_parity.rs"]
+mod avx2_parity;
+use avx2_parity::{assert_vec_parity, close, pseudo_rand};
+
 // ── Tolerance constants ────────────────────────────────────────────────
 
 /// Absolute tolerance for element-wise kernels (embedding).
@@ -28,38 +32,6 @@ const POOL_REL_TOL: f32 = 1e-5;
 const CONV_ABS_TOL: f32 = 1e-4;
 /// Relative tolerance for convolution kernels.
 const CONV_REL_TOL: f32 = 1e-4;
-
-// ── Helpers ────────────────────────────────────────────────────────────
-
-fn close(a: f32, b: f32, abs_tol: f32, rel_tol: f32) -> bool {
-    let diff = (a - b).abs();
-    diff <= abs_tol || diff <= rel_tol * a.abs().max(b.abs())
-}
-
-fn assert_vec_parity(actual: &[f32], expected: &[f32], abs_tol: f32, rel_tol: f32, ctx: &str) {
-    assert_eq!(actual.len(), expected.len(), "{ctx}: length mismatch");
-    for (i, (&a, &e)) in actual.iter().zip(expected.iter()).enumerate() {
-        assert!(
-            close(a, e, abs_tol, rel_tol),
-            "{ctx}[{i}]: scalar={e}, dispatched={a} (diff={}, abs_tol={abs_tol}, rel_tol={rel_tol})",
-            (a - e).abs()
-        );
-    }
-}
-
-/// Deterministic pseudo-random f32 values in [-1, 1] for reproducibility.
-fn pseudo_rand(len: usize, seed: u64) -> Vec<f32> {
-    let mut state = seed;
-    (0..len)
-        .map(|_| {
-            // xorshift64
-            state ^= state << 13;
-            state ^= state >> 7;
-            state ^= state << 17;
-            (state as f32 / u64::MAX as f32) * 2.0 - 1.0
-        })
-        .collect()
-}
 
 /// Deterministic pseudo-random positive f32 values in [0.1, 1.1].
 fn pseudo_rand_positive(len: usize, seed: u64) -> Vec<f32> {
