+++ crates/bitnet-kernels/tests/avx2_matmul_cache_parity_tests.rs
@@ -14,6 +14,10 @@
 use bitnet_kernels::cpu::cache_matmul;
 use bitnet_kernels::cpu::matrix_ops::{self, MatmulConfig};
 
+#[path = "common/avx2_parity.rs"]
+mod avx2_parity;
+use avx2_parity::{assert_vec_parity, pseudo_rand};
+
 // ── Tolerance constants ────────────────────────────────────────────────
 //
 // FMA vs multiply-then-add can differ by up to ~0.5 ULP per op; with
@@ -29,38 +33,6 @@ const MATMUL_REL_TOL: f32 = 1e-3;
 const ELEM_ABS_TOL: f32 = 1e-6;
 /// Relative tolerance for element-wise kernels.
 const ELEM_REL_TOL: f32 = 1e-5;
-
-// ── Helpers ────────────────────────────────────────────────────────────
-
-fn close(a: f32, b: f32, abs_tol: f32, rel_tol: f32) -> bool {
-    let diff = (a - b).abs();
-    diff <= abs_tol || diff <= rel_tol * a.abs().max(b.abs())
-}
-
-fn assert_vec_parity(actual: &[f32], expected: &[f32], abs_tol: f32, rel_tol: f32, ctx: &str) {
-    assert_eq!(actual.len(), expected.len(), "{ctx}: length mismatch");
-    for (i, (&a, &e)) in actual.iter().zip(expected.iter()).enumerate() {
-        assert!(
-            close(a, e, abs_tol, rel_tol),
-            "{ctx}[{i}]: scalar={e}, dispatched={a} (diff={}, abs_tol={abs_tol}, rel_tol={rel_tol})",
-            (a - e).abs()
-        );
-    }
-}
-
-/// Deterministic pseudo-random f32 values in [-1, 1] for reproducibility.
-fn pseudo_rand(len: usize, seed: u64) -> Vec<f32> {
-    let mut state = seed;
-    (0..len)
-        .map(|_| {
-            // xorshift64
-            state ^= state << 13;
-            state ^= state >> 7;
-            state ^= state << 17;
-            (state as f32 / u64::MAX as f32) * 2.0 - 1.0
-        })
-        .collect()
-}
 
 // ── Naïve reference implementations ───────────────────────────────────
