+++ crates/bitnet-kernels/tests/avx2_rope_parity_tests.rs
@@ -30,6 +30,10 @@ use bitnet_kernels::cpu::simd_rope_extended::{
     inverse_rope_rotary_half as ext_inv_rotary_half,
 };
 
+#[path = "common/avx2_parity.rs"]
+mod avx2_parity;
+use avx2_parity::{assert_vec_parity, pseudo_rand};
+
 // ── Tolerance constants ────────────────────────────────────────────────
 //
 // RoPE uses exact sin/cos so AVX2 vs scalar should agree tightly.
@@ -38,38 +42,6 @@ use bitnet_kernels::cpu::simd_rope_extended::{
 const ROPE_ABS_TOL: f32 = 1e-5;
 /// Relative tolerance for RoPE parity checks.
 const ROPE_REL_TOL: f32 = 1e-5;
-
-// ── Helpers ────────────────────────────────────────────────────────────
-
-fn close(a: f32, b: f32, abs_tol: f32, rel_tol: f32) -> bool {
-    let diff = (a - b).abs();
-    diff <= abs_tol || diff <= rel_tol * a.abs().max(b.abs())
-}
-
-fn assert_vec_parity(actual: &[f32], expected: &[f32], abs_tol: f32, rel_tol: f32, ctx: &str) {
-    assert_eq!(actual.len(), expected.len(), "{ctx}: length mismatch");
-    for (i, (&a, &e)) in actual.iter().zip(expected.iter()).enumerate() {
-        assert!(
-            close(a, e, abs_tol, rel_tol),
-            "{ctx}[{i}]: scalar={e}, dispatched={a} (diff={}, abs_tol={abs_tol}, rel_tol={rel_tol})",
-            (a - e).abs()
-        );
-    }
-}
-
-/// Deterministic pseudo-random f32 values in [-1, 1] for reproducibility.
-fn pseudo_rand(len: usize, seed: u64) -> Vec<f32> {
-    let mut state = seed;
-    (0..len)
-        .map(|_| {
-            // xorshift64
-            state ^= state << 13;
-            state ^= state >> 7;
-            state ^= state << 17;
-            (state as f32 / u64::MAX as f32) * 2.0 - 1.0
-        })
-        .collect()
-}
 
 // ════════════════════════════════════════════════════════════════════════
 // 1. BASIC ROPE — cpu::rope
